Move V3Order to own file
